Xlog encryption bugfix: offset calculation was off on TLI change
The Max comparison for dec_off assumed that the key starts in the same
TLI as the current buffer. This is not true if we are switching TLI
during the current server run. Setting this offset incorrectly results
in an overflow and memory corruption later.

Fix: take TLI into account