Draft: supporting multiple WAL key changes
Previously the WAL encryption code assumed that key changes are
relatively rare, and we only ever have to write with the latest
available WAL key.

This is not true, as shown by the 059 TAP test included in this commit:
in replication setups a rejoining node receives the entire last WAL
segment from the current primary.
If the rejoining node already had key changes in this last segment,
rewriting the entire segment with the latest key, without invalidating
these keys is incorrect: later the reading code tries to decrypt the
segment also using these keys, and of course fails.

Invalidating the keys is also a bad idea, as crashes during the segment
rewrite could result in a server not being able to start up at all.

The solution introduced in this commit instead reuses the exisitng keys
for the range they were already used, and uses a new key only for
additional writes.
As the range for the last used key is still open, its endpoint is
detected during reading: as the server tries to read the already
existing WAL first, we can remember this position and use it during the
first write.

This is a draft, as it still has issues. The description of these is in
TODO comments in the xlog_smgr.