Use a single argument for archive command
In the decrypting archivation wrapper.

Instead of having all of the arguments to the archive command separately
and trying to figure out which one should be replaced by the path to the
decrypted WAL segment, we take a single argument and do % parameter
replacement similar to what postgres does with archive_command.

This also mean that we can simplify by using system() instead of exec().