PG-1870 Load pg_tde and setup key in TAP tests
This makes sure pg_tde is loaded and keys are setup when running
postgresql TAP suite. No TDE features are enabled at this point.

When running with meson the setup is sped up by creating a template for
the key files before the first test is run.

In make this setup will currently be done once per test file instead of
once for the whole suite.